+++ Marinel/Data/Entities/InventoryItem.cs
@@ -10,8 +10,8 @@ namespace SchoolDraftWebsite.Data.Entities
         public string Name { get; set; }
         public int InventoryTypeId { get; set; }
         public int Quantity { get; set; }
-        public long CostPerUnit { get; set; }
-        public long SellingPrice { get; set; }
+        public decimal CostPerUnit { get; set; }
+        public decimal SellingPrice { get; set; }
         public virtual InventoryType InventoryType { get; set; }
     }
 }
